# Optimize param validation through `get_param(...) -> Option<Out>`