--- a/crates/bevy_ecs/src/reflect/entity_commands.rs
+++ b/crates/bevy_ecs/src/reflect/entity_commands.rs
@@ -398,7 +398,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands.spawn_empty().id();
         let entity2 = commands.spawn_empty().id();
@@ -435,7 +435,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands.spawn_empty().id();
 
@@ -465,7 +465,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands.spawn(ComponentA(0)).id();
 
@@ -494,7 +494,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands.spawn(ComponentA(0)).id();
 
@@ -523,7 +523,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands.spawn_empty().id();
         let bundle = Box::new(BundleA {
@@ -553,7 +553,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands.spawn_empty().id();
         let bundle = Box::new(BundleA {
@@ -583,7 +583,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands
             .spawn(BundleA {
@@ -621,7 +621,7 @@ mod tests {
         world.insert_resource(type_registry);
 
         let mut system_state: SystemState<Commands> = SystemState::new(&mut world);
-        let mut commands = system_state.get_mut(&mut world);
+        let mut commands = system_state.get_mut(&mut world).unwrap();
 
         let entity = commands
             .spawn(BundleA {