--- a/crates/bevy_ecs/src/system/adapter_system.rs
+++ b/crates/bevy_ecs/src/system/adapter_system.rs
@@ -58,8 +58,8 @@ pub trait Adapt<S: System>: Send + Sync + 'static {
     fn adapt(
         &mut self,
         input: <Self::In as SystemInput>::Inner<'_>,
-        run_system: impl FnOnce(SystemIn<'_, S>) -> S::Out,
-    ) -> Self::Out;
+        run_system: impl FnOnce(SystemIn<'_, S>) -> Option<S::Out>,
+    ) -> Option<Self::Out>;
 }
 
 /// An [`IntoSystem`] creating an instance of [`AdapterSystem`].
@@ -155,7 +155,7 @@ where
         &mut self,
         input: SystemIn<'_, Self>,
         world: UnsafeWorldCell,
-    ) -> Self::Out {
+    ) -> Option<Self::Out> {
         // SAFETY: `system.run_unsafe` has the same invariants as `self.run_unsafe`.
         self.func.adapt(input, |input| unsafe {
             self.system.run_unsafe(input, world)
@@ -163,7 +163,11 @@ where
     }
 
     #[inline]
-    fn run(&mut self, input: SystemIn<'_, Self>, world: &mut crate::prelude::World) -> Self::Out {
+    fn run(
+        &mut self,
+        input: SystemIn<'_, Self>,
+        world: &mut crate::prelude::World,
+    ) -> Option<Self::Out> {
         self.func
             .adapt(input, |input| self.system.run(input, world))
     }
@@ -176,11 +180,6 @@ where
     #[inline]
     fn queue_deferred(&mut self, world: crate::world::DeferredWorld) {
         self.system.queue_deferred(world);
-    }
-
-    #[inline]
-    unsafe fn validate_param_unsafe(&self, world: UnsafeWorldCell) -> bool {
-        self.system.validate_param_unsafe(world)
     }
 
     fn initialize(&mut self, world: &mut crate::prelude::World) {
@@ -228,8 +227,8 @@ where
     fn adapt(
         &mut self,
         input: <Self::In as SystemInput>::Inner<'_>,
-        run_system: impl FnOnce(SystemIn<'_, S>) -> S::Out,
-    ) -> Out {
-        self(run_system(input))
+        run_system: impl FnOnce(SystemIn<'_, S>) -> Option<S::Out>,
+    ) -> Option<Out> {
+        Some(self(run_system(input)?))
     }
 }