--- a/crates/bevy_ecs/src/system/combinator.rs
+++ b/crates/bevy_ecs/src/system/combinator.rs
@@ -101,9 +101,9 @@ pub trait Combine<A: System, B: System> {
     /// See the trait-level docs for [`Combine`] for an example implementation.
     fn combine(
         input: <Self::In as SystemInput>::Inner<'_>,
-        a: impl FnOnce(SystemIn<'_, A>) -> A::Out,
-        b: impl FnOnce(SystemIn<'_, B>) -> B::Out,
-    ) -> Self::Out;
+        a: impl FnOnce(SystemIn<'_, A>) -> Option<A::Out>,
+        b: impl FnOnce(SystemIn<'_, B>) -> Option<B::Out>,
+    ) -> Option<Self::Out>;
 }
 
 /// A [`System`] defined by combining two other systems.
@@ -171,7 +171,7 @@ where
         &mut self,
         input: SystemIn<'_, Self>,
         world: UnsafeWorldCell,
-    ) -> Self::Out {
+    ) -> Option<Self::Out> {
         Func::combine(
             input,
             // SAFETY: The world accesses for both underlying systems have been registered,
@@ -186,7 +186,7 @@ where
         )
     }
 
-    fn run(&mut self, input: SystemIn<'_, Self>, world: &mut World) -> Self::Out {
+    fn run(&mut self, input: SystemIn<'_, Self>, world: &mut World) -> Option<Self::Out> {
         let world = world.as_unsafe_world_cell();
         Func::combine(
             input,
@@ -209,11 +209,6 @@ where
     fn queue_deferred(&mut self, mut world: crate::world::DeferredWorld) {
         self.a.queue_deferred(world.reborrow());
         self.b.queue_deferred(world);
-    }
-
-    #[inline]
-    unsafe fn validate_param_unsafe(&self, world: UnsafeWorldCell) -> bool {
-        self.a.validate_param_unsafe(world) && self.b.validate_param_unsafe(world)
     }
 
     fn initialize(&mut self, world: &mut World) {
@@ -410,13 +405,13 @@ where
         &mut self,
         input: SystemIn<'_, Self>,
         world: UnsafeWorldCell,
-    ) -> Self::Out {
-        let value = self.a.run_unsafe(input, world);
+    ) -> Option<Self::Out> {
+        let value = self.a.run_unsafe(input, world)?;
         self.b.run_unsafe(value, world)
     }
 
-    fn run(&mut self, input: SystemIn<'_, Self>, world: &mut World) -> Self::Out {
-        let value = self.a.run(input, world);
+    fn run(&mut self, input: SystemIn<'_, Self>, world: &mut World) -> Option<Self::Out> {
+        let value = self.a.run(input, world)?;
         self.b.run(value, world)
     }
 
@@ -428,14 +423,6 @@ where
     fn queue_deferred(&mut self, mut world: crate::world::DeferredWorld) {
         self.a.queue_deferred(world.reborrow());
         self.b.queue_deferred(world);
-    }
-
-    unsafe fn validate_param_unsafe(&self, world: UnsafeWorldCell) -> bool {
-        self.a.validate_param_unsafe(world) && self.b.validate_param_unsafe(world)
-    }
-
-    fn validate_param(&mut self, world: &World) -> bool {
-        self.a.validate_param(world) && self.b.validate_param(world)
     }
 
     fn initialize(&mut self, world: &mut World) {