--- a/crates/bevy_ecs/src/system/system_registry.rs
+++ b/crates/bevy_ecs/src/system/system_registry.rs
@@ -343,11 +343,9 @@ impl World {
             initialized = true;
         }
 
-        let result = if system.validate_param(self) {
-            Ok(system.run(input, self))
-        } else {
-            Err(RegisteredSystemError::InvalidParams(id))
-        };
+        let result = system
+            .run(input, self)
+            .ok_or(RegisteredSystemError::InvalidParams(id));
 
         // return ownership of system trait object (if entity still exists)
         if let Some(mut entity) = self.get_entity_mut(id.entity) {