@MiniaczQ 2024-10-03
The steps taken to get here look as following:
1. We question whether 1.2 (from #15505) is possible to implement. While lifetimes and type erasure can be worked around, we can't fetch all system parameters, because in case of compound systems (e.g. `a.pipe(b)`), they can have exclusive parameters. While this isn't technically UB, since the parameters will only be used in mutually exclusive contexts, this will possibly lead to multiple mutable/immutable references to the same data existing at once.
2. To get to 1.2, we need proposition 1.1 anyways. It gives better internal ergonomics than current solution. Because it's only applied when running systems, we don't save on tasks in multithreaded executors nor do we retain the original behavior of running only entire compound systems in executors (e.g. `a.pipe(b)` could run `a` but stop at `b` since params were "checked" lazily).
3. We can fix that by reintroducing `validate_param`, but default it's implementation as `get_param().is_some()`. We expect this implementation to be optimized by compiler. Because default implementation breaks for non-send resources, we overwrite it as well as compound resources (`SystemParam` tuples, `ParamSet`, `DynSystemParam`) to rely on their sub-params implementations.

To sum it up, without going the additional step into proposition 1.2 we simply cannot reduce double checks in executors.
The updated idea can take 2 possible forms:
a) `System`s and `SystemParam`s `get_param()` will return optional resource and multiple conflicting references will exist (this may tip miri off)
b) We separate it into `validate_param() -> Option<Accessor>` which can validate the access without constructing the references and `get_param(Accessor)` which will construct the reference.