@MiniaczQ 2024-10-04
Note for @alice-i-cecile, since I've only made groundwork towards 1.2, but not implemented it, this doesn't have a significant impact on performance and we can delay it.
I've had trouble finding all the broken code, but now that I've done it once it should go much quicker