@MiniaczQ 2024-10-05
I found an edge case that asks for change in requirements:
https://discord.com/channels/691052431525675048/749335865876021248/1292140022375514145
I'm re-drafting this until I implement it