@MiniaczQ 2025-02-26
> @MiniaczQ, this looks like it should be reviewable right? There's a _lot_ of merge conflicts now though 😅

It used to be, all the ideas are here, it just affects such basic APIs that everything conflicts.
I don't have time to resolve this right now, there is a good chance that starting from scratch will be faster, since most of the changes can be regexed