@chescock 2025-02-26
The changes look reasonable, but it seems like a rebase would change enough that I should wait before clicking Approve.  

I don't quite follow the motivation.  It seems like the original goal was to avoid a double lookup, but then you determined that wouldn't be possible.  Is the current goal to remove duplication between `Single::get_param` and `Single::validate_param` and consolidate the calls to `try_warn_param` in one place?  Oh, and to change the behavior of `PipeSystem` so that the second system's parameters are only validated after the first system runs?  

The changes to `SystemState::get` seem unfortunate.  That's a lot of `unwrap()`s!  Is the reasoning that cases like `SystemState::<Single>::get` might now fail, and we need a way to express that?  

Do we expect anyone use fallible parameters with `SystemState`?  The advantage in regular systems is that it gives a way to declare a run condition as part of the system definition, but that doesn't apply to `SystemState`.  I'd vote for leaving `SystemState::get` as panicking so that uses who don't use fallible parameters don't need to `unwrap`, and then exposing a `try_get` for users who do.  (Or, if we really want to reduce accidental panics, we could add `trait InfallibleSystemParam: SystemParam` and bound `get` on it.)